Adding gzip family
Removing troff which is now in APEX repo (pending troff2html).
Removing obsolete mkfiles.

+++ sys/src/cmd/gzip/zip.c
@@ -122,7 +122,7 @@ zip(Biobuf *bout, char *file, int stdout)
 	ZipHead *zh;
 	Dir *dir;
 	int64_t off;
-	int fd, err;
+	int64_t fd, err;
 
 	fd = open(file, OREAD);
 	if(fd < 0)
